feat: add redux slice for searchOptions, useAudioVisualization
