+++ src/components/DetailPage/DiaryDetailAudio.js
@@ -1,121 +1,24 @@
-import React, { useState, useRef, useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
 import { EFFECT_MODE, SCRIPT_MODE } from "../../constants";
 import ErrorModal from "../common/ErrorModal";
 import CustomAudioController from "./CustomAudioController";
+import useAudioVisualization from "./useAudioVisualization";
 
 function DiaryDetailAudio({ mode, diary }) {
+  const { canvasRef, audioElementRef, startDrawingAudio, stopDrawingAudio } =
+    useAudioVisualization({ audioURL: diary?.audio });
+
   const [script, setScript] = useState("");
-
   const [errorMessage, setErrorMessage] = useState("");
-
-  const canvasRef = useRef(null);
-  const canvasContextRef = useRef(null);
-  const animationFrameRef = useRef(null);
-
-  const audioElementRef = useRef(null);
-  const audioContextRef = useRef(null);
-  const audioSourceRef = useRef(null);
-  const audioAnalyzerRef = useRef(null);
 
   useEffect(() => {
     if (mode === SCRIPT_MODE) {
       setScript(diary.script);
     }
   }, [mode]);
-
-  useEffect(() => {
-    if (!audioElementRef.current) {
-      return;
-    }
-
-    audioElementRef.current.src = diary?.audio;
-
-    return () => {
-      cancelAnimationFrame(animationFrameRef.current);
-    };
-  }, [audioElementRef.current, diary, animationFrameRef.current]);
-
-  function startPlayingAudio() {
-    const audio = audioElementRef.current;
-
-    const audioContext =
-      audioContextRef.current ||
-      new (window.AudioContext || window.webkitAudioContext)();
-
-    audioContextRef.current && audioContext.resume();
-
-    const source =
-      audioSourceRef.current || audioContext.createMediaElementSource(audio);
-
-    const analyzer = audioAnalyzerRef.current || audioContext.createAnalyser();
-
-    analyzer.fftSize = 256;
-
-    source.connect(analyzer);
-    analyzer.connect(audioContext.destination);
-
-    const bufferLength = analyzer.frequencyBinCount;
-    const dataArray = new Uint8Array(bufferLength);
-
-    canvasRef.current.width = 600;
-    canvasRef.current.height = 300;
-
-    const barWidth = canvasRef.current.width / bufferLength;
-    let barHeight;
-    let x = 0;
-
-    canvasContextRef.current = canvasRef.current.getContext("2d");
-
-    function animate() {
-      x = 0;
-      clearCanvas();
-      analyzer.getByteFrequencyData(dataArray);
-
-      for (let i = 0; i < bufferLength; i++) {
-        barHeight = dataArray[i];
-        const red = (i * barHeight) / 20;
-        const green = i * 2;
-        const blue = barHeight / 2;
-
-        canvasContextRef.current.fillStyle = `rgb(${red}, ${green}, ${blue})`;
-        canvasContextRef.current.fillRect(
-          x,
-          canvasRef.current.height - barHeight,
-          barWidth,
-          barHeight
-        );
-        x += barWidth;
-      }
-
-      animationFrameRef.current = requestAnimationFrame(animate);
-    }
-
-    animate();
-
-    audioContextRef.current = audioContext;
-    audioAnalyzerRef.current = analyzer;
-    audioSourceRef.current = source;
-  }
-
-  function clearCanvas() {
-    canvasContextRef?.current.clearRect(
-      0,
-      0,
-      canvasRef?.current.width,
-      canvasRef?.current.height
-    );
-  }
-
-  function stopPlayingAudio() {
-    clearCanvas();
-    cancelAnimationFrame(animationFrameRef.current);
-    audioContextRef.current.suspend();
-    audioSourceRef.current.disconnect();
-    audioAnalyzerRef.current.disconnect();
-  }
 
   return (
     <>
@@ -135,8 +38,8 @@ function DiaryDetailAudio({ mode, diary }) {
         <ControllerWrapper>
           <CustomAudioController
             audioElementRef={audioElementRef}
-            onStart={startPlayingAudio}
-            onStop={stopPlayingAudio}
+            onStart={startDrawingAudio}
+            onStop={stopDrawingAudio}
           />
         </ControllerWrapper>
       </Container>
