+++ src/components/EntriesPage/DiaryGraph.js
@@ -1,24 +1,20 @@
 import { useState, useEffect } from "react";
-import { subDays, lightFormat } from "date-fns";
+import { useSelector } from "react-redux";
 import styled from "styled-components";
 
 import noResultImage from "../../assets/images/noResult.png";
 import { getDiaries } from "../../api/axios";
-import { DATE_FORMAT, SENTIMENT } from "../../constants";
 import StyledLoadingSpinner from "../shared/StyledLoadingSpinner";
 import ErrorModal from "../common/ErrorModal";
 import DiarySearch from "./DiarySearch";
 import Graph from "./Graph";
 
 function DiaryGraph() {
+  const searchOptions = useSelector(({ diary }) => diary.searchOptions);
+
   const [diaries, setDiaries] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
   const [errorMessage, setErrorMessage] = useState("");
-  const [searchOptions, setSearchOptions] = useState({
-    startDate: lightFormat(subDays(new Date(), 14), DATE_FORMAT.YYYY_MM_DD),
-    endDate: lightFormat(new Date(), DATE_FORMAT.YYYY_MM_DD),
-    sentiment: SENTIMENT.ALL,
-  });
 
   useEffect(() => {
     (async () => {
@@ -44,7 +40,7 @@ function DiaryGraph() {
 
   return (
     <>
-      <DiarySearch setSearchOptions={setSearchOptions} />
+      <DiarySearch />
       {isLoading && (
         <LoadingWrapper>
           <StyledLoadingSpinner />
