+++ src/features/store.js
@@ -4,14 +4,17 @@ import storage from "redux-persist/lib/storage";
 import thunk from "redux-thunk";
 
 import userReducer from "./userSlice";
+import diaryReducer from "./diarySlice";
 
 const reducers = combineReducers({
   user: userReducer,
+  diary: diaryReducer,
 });
 
 const persistConfig = {
   key: "root",
   storage,
+  blacklist: ["diary"],
 };
 
 const persistedReducer = persistReducer(persistConfig, reducers);
